Remove redundant caching trait
